# docs/api/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Renting Room API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/contracts/{id}/occupants/{occupantId}/end": {
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Ghi nhận người ở rời đi",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "occupantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndOccupancyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/contracts/{id}/vehicles/{vehicleId}/end": {
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Kết thúc đăng ký xe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndVehicleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/exports/renters": {
      "post": {
        "tags": [
          "Exports"
        ],
        "summary": "Xuất Excel danh sách người thuê — lọc theo khu, tầng, nhóm phòng, phòng, khoảng ngày",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRentersQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/renters": {
      "get": {
        "tags": [
          "Renters"
        ],
        "summary": "Danh sách / tìm người thuê theo tên (không dấu) / SĐT. Tìm theo số giấy tờ: POST /renters/search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Renters"
        ],
        "summary": "Tạo hồ sơ người thuê (409 nếu số giấy tờ đã có — dùng lại hồ sơ cũ)",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Khóa duy nhất cho thao tác (VD UUID). Gửi lại cùng khóa ⇒ nhận lại kết quả cũ, không thực thi lần 2.",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 8,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenterInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/renters/search": {
      "post": {
        "tags": [
          "Renters"
        ],
        "summary": "Tìm người thuê, gồm số giấy tờ khớp chính xác (số giấy tờ gửi trong body, không đặt trên URL)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRentersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/renters/{id}": {
      "get": {
        "tags": [
          "Renters"
        ],
        "summary": "Chi tiết người thuê",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Renters"
        ],
        "summary": "Sửa hồ sơ (idNumber = null ⇒ giữ số giấy tờ cũ)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRenterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/renters/{id}/reveal-id-number": {
      "post": {
        "tags": [
          "Renters"
        ],
        "summary": "Xem số giấy tờ đầy đủ (có ghi audit; giới hạn theo user)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/renters/{id}/anonymize": {
      "post": {
        "tags": [
          "Renters"
        ],
        "summary": "Ẩn danh ngay (người thuê yêu cầu xóa dữ liệu / xóa sớm) — chỉ chủ trọ, lý do bắt buộc, KHÔNG đảo ngược",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnonymizeRenterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/rooms/{id}/maintenance/end": {
      "post": {
        "tags": [
          "Rooms"
        ],
        "summary": "Kết thúc bảo trì",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnonymizeRenterRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EndOccupancyRequest": {
        "type": "object",
        "properties": {
          "moveOutDate": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "EndVehicleRequest": {
        "type": "object",
        "properties": {
          "endDate": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "ExportRentersQuery": {
        "type": "object",
        "properties": {
          "propertyIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "floors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "roomGroupIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "roomIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "fromDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "toDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "layout": {
            "$ref": "#/components/schemas/RenterExportLayout"
          },
          "includeSensitive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Gender": {
        "enum": [
          "Male",
          "Female",
          "Other"
        ],
        "type": "string"
      },
...
      "IdDocumentType": {
        "enum": [
          "CitizenId",
          "LegacyId",
          "Passport"
        ],
        "type": "string"
      },
...
      "RenterExportLayout": {
        "enum": [
          "SheetPerProperty",
          "SheetPerFloor",
          "SingleSheet"
        ],
        "type": "string"
      },
      "RenterInput": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "nationality": {
            "type": "string",
            "nullable": true
          },
          "idType": {
            "$ref": "#/components/schemas/IdDocumentType"
          },
          "idNumber": {
            "type": "string",
            "nullable": true
          },
          "idIssueDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "idIssuePlace": {
            "type": "string",
            "nullable": true
          },
          "permanentAddress": {
            "type": "string",
            "nullable": true
          },
          "occupation": {
            "type": "string",
            "nullable": true
          },
          "workplace": {
            "type": "string",
            "nullable": true
          },
          "emergencyContactName": {
            "type": "string",
            "nullable": true
          },
          "emergencyContactPhone": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SearchRentersRequest": {
        "type": "object",
        "properties": {
          "q": {
            "type": "string",
            "nullable": true
          },
          "idNumber": {
            "type": "string",
            "nullable": true
          },
          "idType": {
            "$ref": "#/components/schemas/IdDocumentType"
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateRenterRequest": {
        "type": "object",
        "properties": {
          "renter": {
            "$ref": "#/components/schemas/RenterInput"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Dán access token (không cần gõ chữ 'Bearer').",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}